IPv4 to IPv6 fragment support:
IPv4 to IPv6 one way fragment is now supported.  Note that ICMP
fragment is not supported because it requires to collect all
the fragmented ICMP packet to re-calculate ICMPv6 checksum
(more precisely, the payload length is needed for the re-calculation).
To avoid having a state information for the re-calculation,
we drop the function.