ICMPv4 unreach src changed:
The source address of ICMPv4 destination unreach with
needfrag code set is changed.  It was using the local address
of the map646 node, but now it uses the destination address
of the original IP packet (or corresponding IPv4 address of
the IPv6 node if the original packet is IPv6).  Unless
otherwise, the tun interface doesn't forward the packet.
It seems if the packet source has a local address of the map646
node, the packet sent to the tun interface just dropped.